+++ packages/blockera/php/Routes/api.php
@@ -30,8 +30,7 @@ try {
     $routes->post('telemetry/log-error', [ Blockera\Telemetry\Http\Controllers\BugDetectorAndReporterController::class, 'log' ]);
     $routes->post('telemetry/log-error/status', [ Blockera\Telemetry\Http\Controllers\BugDetectorAndReporterController::class, 'status' ]);
 
-
-	$routes->get('theme-json', [ Blockera\Setup\Http\Controllers\Theme\JSONController::class, 'response' ]);
+	blockera_editor_register_routes($routes);
 
 } catch (Exception $exception) {